fix(component): fixes AllProducts bugs
