+++ client/components/location-verification.js
@@ -1,9 +1,8 @@
 /* eslint-disable no-alert */
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
-import GoogleMapReact from 'google-map-react'
-require('dotenv').config()
 import {updateUser} from '../store/user'
+import UserMap from './UserProfile/UserMap'
 
 class LocationVerification extends Component {
   constructor(props) {
@@ -32,7 +31,7 @@ class LocationVerification extends Component {
       latitude: position.coords.latitude,
       longitude: position.coords.longitude
     })
-    console.log('GEOLOCATION', position)
+    this.addLocation()
   }
   handleLocationError(error) {
     switch (error.code) {
@@ -54,19 +53,21 @@ class LocationVerification extends Component {
   }
   addLocation() {
     if (this.state.latitude && this.state.longitude) {
-      updateUser({
-        ...this.props.user,
-        coordinates: [this.state.latitude, this.state.longitude]
-      })
+      this.props.updateUser(
+        {coordinates: [this.state.latitude, this.state.longitude]},
+        this.props.user.id
+      )
     }
   }
 
   render() {
     return (
       <div>
-        <button onClick={this.getLocation}>Verify location</button>
-        {/* <p>Latitude: {this.state.latitude}</p>
-                <p>Longitude: {this.state.longitude}</p> */}
+        <button type="button" onClick={this.getLocation}>
+          Verify location
+        </button>
+        <p>Latitude: {this.state.latitude}</p>
+        <p>Longitude: {this.state.longitude}</p>
         <UserMap
           userLat={this.state.latitude}
           userLong={this.state.longitude}
@@ -84,40 +85,8 @@ const mapState = state => {
 
 const mapDispatch = dispatch => {
   return {
-    updateUser: userInfo => dispatch(updateUser(userInfo))
+    updateUser: (userInfo, userId) => dispatch(updateUser(userInfo, userId))
   }
 }
 
 export default connect(mapState, mapDispatch)(LocationVerification)
-
-export const UserMap = props => {
-  let userInfo = {
-    center: {
-      lat: props.userLat,
-      lng: props.userLong
-    },
-    zoom: 14
-  }
-
-  const renderMarker = (map, maps) => {
-    let marker = new maps.Marker({
-      position: userInfo.center,
-      map,
-      title: 'User Location'
-    })
-    return marker
-  }
-  return (
-    <div className="user-map" style={{height: '600px', width: '600px'}}>
-      <GoogleMapReact
-        bootstrapURLKeys={{key: process.env.API_KEY}}
-        center={userInfo.center}
-        defaultZoom={userInfo.zoom}
-        yesIWantToUseGoogleMapApiInternals={true}
-        onGoogleApiLoaded={({map, maps}) => {
-          renderMarker(map, maps)
-        }}
-      />
-    </div>
-  )
-}
