+++ client/routes.js
@@ -9,6 +9,7 @@ import {
   LocationVerification,
   Welcome
 } from './components'
+// import {AllProducts} from './components/UserProfile'
 import {me} from './store'
 
 /**
@@ -38,6 +39,7 @@ class Routes extends Component {
           <Switch>
             {/* Routes placed here are only available after logging in */}
             <Route path="/home" component={UserHome} />
+            {/* <Route path="/all-books" component={AllProducts} /> */}
             <Route
               path="/location-verification"
               component={LocationVerification}
